api/requirements.txt
Django==2.2.10
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
importlib-metadata==3.3.0
isort==5.6.4
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.0
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
